Deployment tests finished and configurations adjusted
