+++ samples/spring-webflux-security-xsuaa-usage/pom.xml
@@ -75,6 +75,11 @@
             <artifactId>java-security-test</artifactId>
             <scope>test</scope>
         </dependency>
+        <dependency>
+            <groupId>jakarta.servlet</groupId>
+            <artifactId>jakarta.servlet-api</artifactId>
+            <version>6.0.0</version>
+        </dependency>
     </dependencies>
 
     <build>
