+++ samples/spring-webflux-security-xsuaa-usage/src/main/resources/application.yml
@@ -8,3 +8,7 @@ logging.level:
   org.springframework.security: DEBUG # set to ERROR for production setups. 
   org.springframework.webflux: DEBUG      # set to ERROR for production setups.
   sample.spring.webflux.xsuaa: INFO
+spring:
+  main:
+    allow-bean-definition-overriding: true
+
